+++ registry/remote/repository.go
@@ -35,8 +35,6 @@ import (
 	"oras.land/oras-go/v2/content"
 	"oras.land/oras-go/v2/errdef"
 	"oras.land/oras-go/v2/internal/cas"
-	"oras.land/oras-go/v2/internal/container/set"
-	"oras.land/oras-go/v2/internal/descriptor"
 	"oras.land/oras-go/v2/internal/httputil"
 	"oras.land/oras-go/v2/internal/ioutil"
 	"oras.land/oras-go/v2/internal/registryutil"
@@ -53,44 +51,7 @@ const (
 	// See https://docs.docker.com/registry/spec/api/#digest-header
 	// See https://github.com/opencontainers/distribution-spec/blob/main/spec.md#pull
 	dockerContentDigestHeader = "Docker-Content-Digest"
-	// zeroDigest represents a digest that consists of zeros. zeroDigest is used
-	// for pinging Referrers API.
-	zeroDigest = "sha256:0000000000000000000000000000000000000000000000000000000000000000"
 )
-
-// referrersState represents the state of Referrers API.
-type referrersState = int32
-
-const (
-	// referrersStateUnknown represents an unknown state of Referrers API.
-	referrersStateUnknown referrersState = iota
-	// referrersStateSupported represents that the repository is known to
-	// support Referrers API
-	referrersStateSupported
-	// referrersStateUnsupported represents that the repository is known to
-	// not support Referrers API
-	referrersStateUnsupported
-)
-
-// referrerOperation represents an operation on a referrer.
-type referrerOperation = int32
-
-const (
-	// referrersStateUnknown represents an addition operation on a referrer.
-	referrerOperationAdd referrerOperation = iota
-	// referrersStateUnknown represents a removal operation on a referrer.
-	referrerOperationRemove
-)
-
-// referrerChange represents a change on a referrer.
-type referrerChange struct {
-	referrer  ocispec.Descriptor
-	operation referrerOperation
-}
-
-// ErrReferrersCapabilityAlreadySet is returned by SetReferrersCapability()
-// when the Referrers API capability has been already set.
-var ErrReferrersCapabilityAlreadySet = errors.New("referrers capability cannot be changed once set")
 
 // Client is an interface for a HTTP client.
 type Client interface {
@@ -145,7 +106,7 @@ type Repository struct {
 	// default: referrersStateUnknown
 	referrersState referrersState
 
-	// referrersPingLock locks the pingReferrersAPI() method and allows only
+	// referrersPingLock locks the pingReferrers() method and allows only
 	// one go-routine to send the request.
 	referrersPingLock sync.Mutex
 
@@ -589,49 +550,6 @@ func (r *Repository) pingReferrers(ctx context.Context) (bool, error) {
 	default:
 		return false, errutil.ParseErrorResponse(resp)
 	}
-}
-
-// buildReferrersTag builds the referrers tag for the given manifest descriptor.
-// Format: <algorithm>-<digest>
-// Reference: https://github.com/opencontainers/distribution-spec/blob/v1.1.0-rc1/spec.md#unavailable-referrers-api
-func buildReferrersTag(desc ocispec.Descriptor) string {
-	alg := desc.Digest.Algorithm().String()
-	encoded := desc.Digest.Encoded()
-	return alg + "-" + encoded
-}
-
-// isReferrersFilterApplied checks annotations to see if requested is in the
-// applied filter list.
-func isReferrersFilterApplied(annotations map[string]string, requested string) bool {
-	applied := annotations[ocispec.AnnotationReferrersFiltersApplied]
-	if applied == "" || requested == "" {
-		return false
-	}
-	filters := strings.Split(applied, ",")
-	for _, f := range filters {
-		if f == requested {
-			return true
-		}
-	}
-	return false
-}
-
-// filterReferrers filters a slice of referrers by artifactType in place.
-// The returned slice contains matching referrers.
-func filterReferrers(refs []ocispec.Descriptor, artifactType string) []ocispec.Descriptor {
-	if artifactType == "" {
-		return refs
-	}
-	var j int
-	for i, ref := range refs {
-		if ref.ArtifactType == artifactType {
-			if i != j {
-				refs[j] = ref
-			}
-			j++
-		}
-	}
-	return refs[:j]
 }
 
 // delete removes the content identified by the descriptor in the entity "blobs"
@@ -1331,73 +1249,6 @@ func (s *manifestStore) updateReferrersIndex(ctx context.Context, desc, subject 
 		return nil
 	}
 	return nil
-}
-
-// errNoReferrerUpdate is returned by applyReferrerChanges() when there
-// is no any referrer update.
-var errNoReferrerUpdate = errors.New("no referrer update")
-
-// applyReferrerChanges applies referrerChanges on referrers and returns the
-// updated referrers.
-func applyReferrerChanges(referrers []ocispec.Descriptor, referrerChanges []referrerChange) ([]ocispec.Descriptor, error) {
-	referrersSet := make(set.Set[descriptor.Descriptor], len(referrers))
-	for _, r := range referrers {
-		key := descriptor.FromOCI(r)
-		referrersSet.Add(key)
-	}
-
-	var referrersToAdd []ocispec.Descriptor
-	referrersToRemove := make(set.Set[descriptor.Descriptor], len(referrers))
-	for _, change := range referrerChanges {
-		key := descriptor.FromOCI(change.referrer)
-		switch change.operation {
-		case referrerOperationAdd:
-			if !referrersSet.Contains(key) {
-				// add distinct referrers
-				referrersSet.Add(key)
-				referrersToAdd = append(referrersToAdd, change.referrer)
-			}
-		case referrerOperationRemove:
-			// log distinct referrers to remove
-			referrersToRemove.Add(key)
-		}
-	}
-	if len(referrersToAdd) == len(referrersToRemove) {
-		// the changes can be offset when the items in referrersToAdd and
-		// referrersToRemove are the same
-		canOffset := true
-		for _, r := range referrersToAdd {
-			key := descriptor.FromOCI(r)
-			if !referrersToRemove.Contains(key) {
-				canOffset = false
-				break
-			}
-		}
-		if canOffset {
-			return nil, errNoReferrerUpdate
-		}
-	}
-
-	var referrersUpdated bool
-	var updatedReferrers []ocispec.Descriptor
-	if len(referrersToAdd) > 0 {
-		referrers = append(referrers, referrersToAdd...)
-		referrersUpdated = true
-	}
-	for _, r := range referrers {
-		key := descriptor.FromOCI(r)
-		if referrersToRemove.Contains(key) {
-			// exclude the referrers that are in the removal set
-			referrersUpdated = true
-		} else {
-			updatedReferrers = append(updatedReferrers, r)
-		}
-	}
-	if !referrersUpdated {
-		return nil, errNoReferrerUpdate
-	}
-
-	return updatedReferrers, nil
 }
 
 // ParseReference parses a reference to a fully qualified reference.
